change: Tipo dos identificadores da entidade Atendimento

+++ wbhealth/src/main/java/br/com/dbc/wbhealth/model/entity/Atendimento.java
@@ -20,7 +20,7 @@ public class Atendimento {
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "ATENDIMENTO_SEQ")
     @SequenceGenerator(name = "ATENDIMENTO_SEQ", sequenceName = "SEQ_ATENDIMENTO", allocationSize = 1)
     @Column(name = "id_atendimento")
-    private Long idAtendimento;
+    private Integer idAtendimento;
 
 //    @OneToOne(fetch = FetchType.LAZY)
 //    @JoinColumn(name = "id_hospital", referencedColumnName = "id_hospital")
@@ -34,14 +34,14 @@ public class Atendimento {
 //    @JoinColumn(name = "id_paciente", referencedColumnName = "id_paciente")
 //    private Paciente idPaciente;
     @Column(name = "id_paciente")
-    private Long idPaciente;
+    private Integer idPaciente;
 
 //    @OneToOne(fetch = FetchType.LAZY)
 //    @JoinColumn(name = "id_medico", referencedColumnName = "id_medico", table = "medico")
 //    private Medico idMedico;
 
     @Column(name = "id_medico")
-    private Long idMedico;
+    private Integer idMedico;
 
     @Column(name = "data_atendimento")
     private LocalDate dataAtendimento;
